requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.0.0
httpx>=0.25.0
python-dateutil>=2.8.2
# Nota: fastapi e uvicorn requerem pydantic que precisa compilar no Python 3.15
# Use Python 3.11 ou 3.12, ou instale Visual C++ Build Tools
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

